refactor: '/items'페이지를 위한 기존 코드 리팩토링
