+++ components/items/AllItemsSection.tsx
@@ -1,71 +1,66 @@
 import { useEffect, useState } from 'react';
-import { getProducts } from '../lib/api';
-import ItemCard from '@components/ItemCard';
-import '../style/Items.css';
+import { getProducts } from '@lib/api';
+import ItemCard from '@components/items/ItemCard';
+import '@styles/Items.module.css';
 import SearchIcon from '@public/svgs/ic_search.svg';
 import {
   Product,
   ProductListResponse,
   ProductOrderBy,
-} from './types/productTypes';
+} from '@components/types/productTypes';
 import Link from 'next/link';
 
-const getPageSize = () => {
-  const width = window.innerWidth;
+const getPageSize = (width: number): number => {
   if (width < 768) {
-    // Mobile viewport
-    return 4;
+    return 4; // Mobile viewport
   } else if (width < 1280) {
-    // Tablet viewport
-    return 6;
+    return 6; // Tablet viewport
   } else {
-    // Desktop viewport
-    return 10;
+    return 10; // Desktop viewport
   }
 };
 
 function AllItemsSection() {
   const [orderBy, setOrderBy] = useState<ProductOrderBy>('recent');
   const [page, setPage] = useState(1);
-  const [pageSize, setPageSize] = useState(getPageSize());
+
+  // 초기값 설정에서 window.innerWidth 사용
+  const [pageSize, setPageSize] = useState(
+    typeof window !== 'undefined' ? getPageSize(window.innerWidth) : 10,
+  );
+
   const [itemList, setItemList] = useState<Product[]>([]);
   const [isDropdownVisible, setIsDropdownVisible] = useState(false);
   const [totalPageNum, setTotalPageNum] = useState(1);
 
-  const fetchSortedData = async ({
-    orderBy,
-    page,
-    pageSize,
-  }: {
+  const fetchSortedData = async (params: {
     orderBy: ProductOrderBy;
     page: number;
     pageSize: number;
   }) => {
     try {
-      const products: ProductListResponse = await getProducts({
-        orderBy,
-        page,
-        pageSize,
-      });
+      const products: ProductListResponse = await getProducts(params);
       setItemList(products.list);
-      setTotalPageNum(Math.ceil(products.totalCount / pageSize));
+      setTotalPageNum(Math.ceil(products.totalCount / params.pageSize));
     } catch (error) {
       throw new Error('정보를 불러오는데 실패했습니다.');
     }
   };
 
   useEffect(() => {
-    const handleFixsize = () => {
-      setPageSize(getPageSize());
+    const handleResize = () => {
+      setPageSize(getPageSize(window.innerWidth));
     };
 
-    // 화면 크기 변경할 때마다 pageSize를 다시 계산해 넣음
-    window.addEventListener('resize', handleFixsize);
+    // 화면 크기 변경할 때마다 pageSize를 다시 계산
+    window.addEventListener('resize', handleResize);
+
+    // 데이터 가져오기
     fetchSortedData({ orderBy, page, pageSize });
 
     // Cleanup function
     return () => {
-      window.removeEventListener('resize', handleFixsize);
+      window.removeEventListener('resize', handleResize);
     };
   }, [orderBy, page, pageSize]);
