+++ components/items/BestItemsSection.tsx
@@ -1,30 +1,28 @@
 import { useEffect, useState } from 'react';
 import { getProducts } from '@lib/api';
-import ItemCard from '@components/ItemCard';
-import '../style/Items.css';
+import ItemCard from '@/components/items/ItemCard';
+import '@styles/Items.module.css';
 import {
   Product,
   ProductListResponse,
   ProductOrderBy,
-} from './types/productTypes';
+} from '../types/productTypes';
 
-const getPageSize = () => {
-  const width = window.innerWidth;
+const getPageSize = (width: number): number => {
   if (width < 768) {
-    // Mobile viewport
-    return 1;
+    return 1; // Mobile viewport
   } else if (width < 1280) {
-    // Tablet viewport
-    return 2;
+    return 2; // Tablet viewport
   } else {
-    // Desktop viewport
-    return 4;
+    return 3; // Desktop viewport
   }
 };
 
 function BestItemSection() {
   const [itemList, setItemList] = useState<Product[]>([]);
-  const [pageSize, setPageSize] = useState(getPageSize());
+  const [pageSize, setPageSize] = useState(
+    () => (typeof window !== 'undefined' ? getPageSize(window.innerWidth) : 3), // 초기값
+  );
 
   const fetchSortedData = async ({
     orderBy,
@@ -41,15 +39,15 @@ function BestItemSection() {
   };
 
   useEffect(() => {
-    const handleFixsize = () => {
-      setPageSize(getPageSize());
+    const handleResize = () => {
+      setPageSize(getPageSize(window.innerWidth));
     };
 
-    window.addEventListener('resize', handleFixsize);
+    window.addEventListener('resize', handleResize);
     fetchSortedData({ orderBy: 'favorite', pageSize });
 
     return () => {
-      window.removeEventListener('resize', handleFixsize);
+      window.removeEventListener('resize', handleResize);
     };
   }, [pageSize]);
